Scope4Mac v0.6.0-mac — Daydream Scope fork for Apple Silicon MPS
Full port of Daydream Scope to run on Apple Silicon with unified memory.
Built for a Daydream hackathon on M2 Max 96GB.

Core MPS port (60+ files):
- Device selection: CUDA → MPS → CPU fallback chain
- Flash attention → SDPA fallback (9 call sites)
- flex_attention → SDPA with is_causal (5 pipelines)
- fp8 → fp16 text encoder (bf16 alternative artifact)
- float64 → float32 everywhere (28 occurrences, 12 files)
- VAE autocast device_type: "cpu" → "mps" (9 sites)
- VAE decode .to("cuda") → .to(latent.device)
- torch.compile: skip inductor on non-CUDA
- RIFE: grid_sample padding_mode='zeros' (MPS unsupported border)
- Unified memory detection via os.sysconf

New turbo4mac pipeline:
- SD-Turbo 1-step img2img + TAESD (25x faster VAE)
- AutoPipeline with prompt caching + output_type="pt"
- 6 FPS real output at 256x256 on M2 Max
- Registered as Scope plugin in pipeline dropdown

UI/UX:
- Frutiger Aero skin (from construct aesthetics-librarian spec)
- Mac OS X Cheetah wallpaper background
- Happy Mac icon + serif italic title
- Chrome tribal wing decorations
- Unified memory gauge in status bar
- 5-minute pipeline load timeout for large models
- Rebranded: Daydream Scope4Mac

Upstream cherry-pick:
- FPS throttle fix from v0.1.8 (PR #701)

+++ src/scope/core/pipelines/wan2_1/components/generator.py
@@ -88,7 +88,7 @@ class WanDiffusionWrapper(torch.nn.Module):
                         torch.arange(max_seq_len),
                         1.0
                         / torch.pow(
-                            theta, torch.arange(0, dim, 2).to(torch.float64).div(dim)
+                            theta, torch.arange(0, dim, 2).to(torch.float32).div(dim)
                         ),
                     )
                     freqs = torch.polar(torch.ones_like(freqs), freqs)
@@ -151,7 +151,7 @@ class WanDiffusionWrapper(torch.nn.Module):
         # use higher precision for calculations
         original_dtype = flow_pred.dtype
         flow_pred, xt, sigmas, timesteps = [
-            x.double().to(flow_pred.device)
+            x.float().to(flow_pred.device)
             for x in [flow_pred, xt, self.scheduler.sigmas, self.scheduler.timesteps]
         ]
 
@@ -177,7 +177,7 @@ class WanDiffusionWrapper(torch.nn.Module):
         # use higher precision for calculations
         original_dtype = x0_pred.dtype
         x0_pred, xt, sigmas, timesteps = [
-            x.double().to(x0_pred.device)
+            x.float().to(x0_pred.device)
             for x in [x0_pred, xt, scheduler.sigmas, scheduler.timesteps]
         ]
         timestep_id = torch.argmin(
