+++ src/scope/core/pipelines/wan2_1/components/scheduler.py
@@ -39,7 +39,7 @@ class SchedulerInterface(ABC):
         # use higher precision for calculations
         original_dtype = x0.dtype
         x0, xt, alphas_cumprod = (
-            x.double().to(x0.device) for x in [x0, xt, self.alphas_cumprod]
+            x.float().to(x0.device) for x in [x0, xt, self.alphas_cumprod]
         )
 
         alpha_prod_t = alphas_cumprod[timestep].reshape(-1, 1, 1, 1)
@@ -62,7 +62,7 @@ class SchedulerInterface(ABC):
         # use higher precision for calculations
         original_dtype = noise.dtype
         noise, xt, alphas_cumprod = (
-            x.double().to(noise.device) for x in [noise, xt, self.alphas_cumprod]
+            x.float().to(noise.device) for x in [noise, xt, self.alphas_cumprod]
         )
         alpha_prod_t = alphas_cumprod[timestep].reshape(-1, 1, 1, 1)
         beta_prod_t = 1 - alpha_prod_t
@@ -88,7 +88,7 @@ class SchedulerInterface(ABC):
         # use higher precision for calculations
         original_dtype = velocity.dtype
         velocity, xt, alphas_cumprod = (
-            x.double().to(velocity.device) for x in [velocity, xt, self.alphas_cumprod]
+            x.float().to(velocity.device) for x in [velocity, xt, self.alphas_cumprod]
         )
         alpha_prod_t = alphas_cumprod[timestep].reshape(-1, 1, 1, 1)
         beta_prod_t = 1 - alpha_prod_t
